visualisator/requirements.txt
flask 
grpcio-tools ~= 1.30
Jinja2 
pytest